+++ src/commands/voice/choose_successor.ts
@@ -2,6 +2,8 @@ import { ApplicationCommandOptionType, ApplicationCommandType } from "discord.js
 import { SlashCommand } from "../../structures/Command";
 import isArrayElement from "../../lib/isArrayElement";
 import AutoVoices from "../../schemas/AutoVoices";
+import CommandChannels from "../../schemas/CommandChannels";
+import isVoiceCommandChannel from "../../lib/isVoiceCommandChannel";
 
 export default new SlashCommand({
   name: 'voice-successor',
@@ -21,10 +23,24 @@ export default new SlashCommand({
   run: async ({ interaction }) => {
     const targetUser = interaction.options.getUser("user");
     const currentVoice = interaction.member.voice.channel;   
+    const currentChannel = interaction.channel;
     const cummandUsed = interaction.member;
 
     const channel_from_db = await AutoVoices.findOne({channel_id: currentVoice.id});
     const successorsArray = channel_from_db.successors;
+    const specialChannelsArray: string[] = await (await CommandChannels.find())
+      .map((e) => {
+        return e.channel_id
+      })
+       
+    if (!isVoiceCommandChannel(currentChannel.id, specialChannelsArray)) {
+      interaction.reply({
+        content: `Используйте специальный канал для войс комманд.`,
+        ephemeral: true
+      })
+
+      return;
+    }
     
     if (!channel_from_db) {
       currentVoice.delete().catch(() => {});
