+++ src/commands/voice/kick-user.ts
@@ -1,6 +1,8 @@
 import { ApplicationCommandOptionType, ApplicationCommandType } from "discord.js";
 import isArrayElement from "../../lib/isArrayElement";
+import isVoiceCommandChannel from "../../lib/isVoiceCommandChannel";
 import AutoVoices from "../../schemas/AutoVoices";
+import CommandChannels from "../../schemas/CommandChannels";
 import { SlashCommand } from "../../structures/Command";
 
 export default new SlashCommand({
@@ -23,9 +25,24 @@ export default new SlashCommand({
     const { guild } = interaction;
     const currentVoice = interaction.member.voice.channel;   
     const cummandUsed = interaction.member;
+    const currentChannel = interaction.channel;
 
     const channel_from_db = await AutoVoices.findOne({channel_id: currentVoice.id});
     const successorsArray = channel_from_db.successors; 
+
+    const specialChannelsArray: string[] = await (await CommandChannels.find())
+      .map((e) => {
+        return e.channel_id
+      })
+    
+    if (!isVoiceCommandChannel(currentChannel.id, specialChannelsArray)) {
+      interaction.reply({
+        content: `Используйте специальный канал для войс комманд.`,
+        ephemeral: true
+      })
+
+      return;
+    }
 
     if (!channel_from_db) {
       currentVoice.delete().catch(() => {});
